update246

+++ controllers/categoryCtrl.js
@@ -13,8 +13,6 @@ const categoryCtrl = {
     },
     createCategory: async (req, res) =>{
         try {
-            // if user have role = 1 ---> admin
-            // only admin can create , delete and update category
             const {name} = req.body;
             const category = await Category.findOne({name})
             if(category) return res.status(400).json({msg: "This category already exists."})
@@ -30,10 +28,10 @@ const categoryCtrl = {
     },
     deleteCategory: async(req, res) =>{
         try {
-            // const products = await Products.findOne({category: req.params.id})
-            // if(products) return res.status(400).json({
-            //     msg: "Please delete all products with a relationship."
-            // })
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({
+                msg: "Please delete all products with a relationship."
+            })
 
             await Category.findByIdAndDelete(req.params.id)
             res.json({msg: "Deleted a Category"})
